// src/main/typescript/autobahn_client/proto/message.ts
export enum MessageType {
  SERVER_STATE = 0,
  SERVER_FORWARD = 1,
  PUBLISH = 2,
  SUBSCRIBE = 3,
  UNSUBSCRIBE = 4,
  HEARTBEAT = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface TopicMessage {
  messageType: MessageType;
  topic: string;
}

export interface PublishMessage {
  messageType: MessageType;
  topic: string;
  payload: Uint8Array;
}

export interface UnsubscribeMessage {
  messageType: MessageType;
  topic: string;
}
// ...
export const TopicMessage = {
  encode(message: TopicMessage, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.messageType !== 0) {
      writer.uint32(8).int32(message.messageType);
    }
    if (message.topic !== "") {
      writer.uint32(18).string(message.topic);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): TopicMessage {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTopicMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.messageType = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.topic = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TopicMessage {
    return {
      messageType: isSet(object.messageType) ? messageTypeFromJSON(object.messageType) : 0,
      topic: isSet(object.topic) ? globalThis.String(object.topic) : "",
    };
  },

  toJSON(message: TopicMessage): unknown {
    const obj: any = {};
    if (message.messageType !== 0) {
      obj.messageType = messageTypeToJSON(message.messageType);
    }
    if (message.topic !== "") {
      obj.topic = message.topic;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TopicMessage>, I>>(base?: I): TopicMessage {
    return TopicMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TopicMessage>, I>>(object: I): TopicMessage {
    const message = createBaseTopicMessage();
    message.messageType = object.messageType ?? 0;
    message.topic = object.topic ?? "";
    return message;
  },
};

function createBasePublishMessage(): PublishMessage {
  return { messageType: 0, topic: "", payload: new Uint8Array(0) };
}

export const PublishMessage = {
  encode(message: PublishMessage, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.messageType !== 0) {
      writer.uint32(8).int32(message.messageType);
    }
    if (message.topic !== "") {
      writer.uint32(18).string(message.topic);
    }
    if (message.payload.length !== 0) {
      writer.uint32(26).bytes(message.payload);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PublishMessage {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePublishMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.messageType = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.topic = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.payload = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PublishMessage {
    return {
      messageType: isSet(object.messageType) ? messageTypeFromJSON(object.messageType) : 0,
      topic: isSet(object.topic) ? globalThis.String(object.topic) : "",
      payload: isSet(object.payload) ? bytesFromBase64(object.payload) : new Uint8Array(0),
    };
  },

  toJSON(message: PublishMessage): unknown {
    const obj: any = {};
    if (message.messageType !== 0) {
      obj.messageType = messageTypeToJSON(message.messageType);
    }
    if (message.topic !== "") {
      obj.topic = message.topic;
    }
    if (message.payload.length !== 0) {
      obj.payload = base64FromBytes(message.payload);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PublishMessage>, I>>(base?: I): PublishMessage {
    return PublishMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PublishMessage>, I>>(object: I): PublishMessage {
    const message = createBasePublishMessage();
    message.messageType = object.messageType ?? 0;
    message.topic = object.topic ?? "";
    message.payload = object.payload ?? new Uint8Array(0);
    return message;
  },
};

function createBaseUnsubscribeMessage(): UnsubscribeMessage {
  return { messageType: 0, topic: "" };
}

export const UnsubscribeMessage = {
  encode(message: UnsubscribeMessage, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.messageType !== 0) {
      writer.uint32(8).int32(message.messageType);
    }
    if (message.topic !== "") {
      writer.uint32(18).string(message.topic);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): UnsubscribeMessage {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUnsubscribeMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.messageType = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.topic = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UnsubscribeMessage {
    return {
      messageType: isSet(object.messageType) ? messageTypeFromJSON(object.messageType) : 0,
      topic: isSet(object.topic) ? globalThis.String(object.topic) : "",
    };
  },

  toJSON(message: UnsubscribeMessage): unknown {
    const obj: any = {};
    if (message.messageType !== 0) {
      obj.messageType = messageTypeToJSON(message.messageType);
    }
    if (message.topic !== "") {
      obj.topic = message.topic;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UnsubscribeMessage>, I>>(base?: I): UnsubscribeMessage {
    return UnsubscribeMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UnsubscribeMessage>, I>>(object: I): UnsubscribeMessage {
    const message = createBaseUnsubscribeMessage();
    message.messageType = object.messageType ?? 0;
    message.topic = object.topic ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };